clean up conver_model examples
