+++ convert_model/__init__.py
@@ -1,2 +1,7 @@
+import tensorflow as tf
+
+keras = tf.keras
+
+
 def red(string: str) -> str:
     return f"\033[91m{string}\033[0m"
